Graphs/requirements.txt
dash==2.16.1
plotly==5.3.1
pandas==1.3.3
psycopg2-binary==2.9.1
Flask==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.20.0